ai/requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
python-dotenv>=1.0.0
anthropic>=0.40.0
